feature: add proxy receiveSlot for detecting more results
We try to choose one more result than we really want; we don't actually return this last value, but
we do consume it in order to tell if we would have had more data.

We implement this via the pg_limit_proxy_slot routine, which we inject into the existing QueryDesc
struct in our ExecutorRun_hook, so we are able to terminate at the point we want, without needing to
care that much about the details of where this tuple is actually going.

In the future, we may want to distinguish different query types here and handle things differently;
for instance, this max_rows limit currently affects *all* query types, including COPY output, INSERT
INTO foo SELECT, etc; if the main point of this restriction is limiting the number of rows returned
the user, then filtering based on the destination slot type could be useful.